adding files

+++ doubly_linked_lists/2-add_dnodeint.c
@@ -13,7 +13,7 @@ dlistint_t *add_dnodeint(dlistint_t **head, const int n)
 	tmp = malloc(sizeof(dlistint_t));
 	if (tmp == NULL)
 		return (NULL);
-	tmp->n = n;
+	tmp->n = (int) n;
 	tmp->prev = NULL;
 
 	if (head != NULL)
